antiwey@gmail.com20190617: 1. 对宠物信息输入页面做输入值的检查，必须有输入才能跳转到下一页 2. 宠物体重改为Double型 3. 宠物信息页面单选项添加默认值

+++ project/app/src/main/java/com/petbus/tj/petbus/middleware/middleware.java
@@ -24,6 +24,6 @@ public interface middleware{
     int get_last_three_record( Map<String,String> record_map );
     int new_record( String time, String petname, String action, String remark, ArrayList<String> record_pic );
     int get_record( int id, StringBuffer time, StringBuffer petname, StringBuffer action, StringBuffer remark, ArrayList<String> record_pic );
-    int newPet(String name, String photoPath, String birth, int weight, int gender, int species);
+    int newPet(String name, String photoPath, String birth, double weight, int gender, int species);
     int getPetInfo(int id, String name, String photoPath, String birth, int weight, int gender, int species);
 }
